Extracted blacklist to its own interface
* Moved WhitelistMode enum to the api package.
