+++ src/main/java/media/xen/tradingcards/api/blacklist/Blacklist.java
@@ -1,0 +1,16 @@
+package media.xen.tradingcards.api.blacklist;
+
+public interface Blacklist<T> {
+    boolean isAllowed(T value);
+
+    void add(T value);
+
+
+    void remove(T value);
+
+    /**
+     * Returns the mode the whitelist is in.
+     * @return
+     */
+    WhitelistMode getMode();
+}
