+++ src/main/java/media/xen/tradingcards/api/blacklist/WhitelistMode.java
@@ -1,4 +1,4 @@
-package media.xen.tradingcards.whitelist;
+package media.xen.tradingcards.api.blacklist;
 /**
  * @author ketelsb
  */
